move loadFixture into the NetworkHelpers class
