+++ v-next/hardhat-network-helpers/src/internal/network-helpers/helpers/reset.ts
@@ -1,15 +1,16 @@
 import type { NumberLike } from "../../../types.js";
+import type { NetworkHelpers } from "../network-helpers.js";
 import type { EthereumProvider } from "@ignored/hardhat-vnext/types/providers";
 
-import { clearSnapshots } from "../../../load-fixture.js";
 import { toNumber } from "../../conversion.js";
 
 export async function reset(
+  networkHelpers: NetworkHelpers,
   provider: EthereumProvider,
   url?: string,
   blockNumber?: NumberLike,
 ): Promise<void> {
-  clearSnapshots();
+  networkHelpers.clearSnapshots();
 
   if (url === undefined) {
     await provider.request({ method: "hardhat_reset", params: [] });
